Factor ModuleResolver out to its own module
